backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.4.2
httpx==0.27.0
xgboost==2.0.3
scikit-learn==1.7.0
pandas==2.2.3
numpy==2.2.6
geopy==2.4.1
pydantic==2.11.7
python-multipart==0.0.9
SQLAlchemy==2.0.38
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
